scripts/translations/requirements.txt
google-cloud-translate
boto3
